requirements.txt
pandas
matplotlib
numpy
google-genai
dotenv
ipykernel
scikit-learn
wordcloud
joblib
gradio